Create logic to compare current DB (xlsm from dropbox) with the latest DB used in bot (CSV) (#30)
Now we can compare the two files. Whenever the DB constructor is
launched, we also need a path to the latest state of the DB used by the
bot (`data/latest_collection.csv`). We then fetch file from server and
compare row-by-row. If we detect a status change, we collect the date
where the coin was collected. This will later allow to fully resolve #23

- Improved logic to re-download the XLM file from Dropbox. When bot
starts, it always downloads, afterwards only when the file was updated
on the remote (etag change)